requirements.txt
amqp==5.1.1
asgiref==3.7.2
autopep8==2.0.2
billiard==4.1.0
celery==5.3.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
defusedxml==0.7.1
Django==4.2.3
django-silk==5.0.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
fonttools==4.41.1
gprof2dot==2022.7.29
idna==3.4
kiwisolver==1.4.4
kombu==5.3.2
matplotlib==3.7.2
numpy==1.25.1
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
